requirements.txt
python-dotenv
fireworks-ai
rich